extracted useUpdatableState to its own file
